// src/services/models/createMemeSwapRequestResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */

export interface CreateMemeSwapRequestResponse {
  readonly id: string;
  readonly tokenAddress: string;
  readonly sender: string;
  readonly remainingTokenAmountAllocation: string;
  readonly expiredBlockNumber: string;
  readonly signature: string;
}